refactor: split into FileWriter and TableWriter classes
