+++ Tests/StorageApiWriterMetadataTest.php
@@ -4,7 +4,7 @@ namespace Keboola\DockerBundle\Tests;
 
 use Keboola\Csv\CsvFile;
 use Keboola\OutputMapping\Exception\InvalidOutputException;
-use Keboola\OutputMapping\Writer\Writer;
+use Keboola\OutputMapping\Writer\TableWriter;
 use Keboola\StorageApi\Client;
 use Keboola\StorageApi\ClientException;
 use Keboola\StorageApi\Metadata;
@@ -133,7 +133,7 @@ class StorageApiWriterMetadataTest extends \PHPUnit_Framework_TestCase
             "configurationId" => "metadata-write-test"
         ];
 
-        $writer = new Writer($this->client, new NullLogger());
+        $writer = new TableWriter($this->client, new NullLogger());
         $tableQueue =  $writer->uploadTables($root . "/upload", $config, $systemMetadata);
         $jobIds = $tableQueue->waitForAll();
         $this->assertCount(1, $jobIds);
@@ -197,7 +197,7 @@ class StorageApiWriterMetadataTest extends \PHPUnit_Framework_TestCase
         ];
         $systemMetadata = ["componentId" => "testComponent"];
 
-        $writer = new Writer($this->client, new NullLogger());
+        $writer = new TableWriter($this->client, new NullLogger());
         $tableQueue =  $writer->uploadTables($root . "/upload", $config, $systemMetadata);
         $this->expectException(InvalidOutputException::class);
         $this->expectExceptionMessage('Failed to load table "in.c-docker-test.table55a": Load error: ' .
@@ -256,7 +256,7 @@ class StorageApiWriterMetadataTest extends \PHPUnit_Framework_TestCase
                 ],
             ],
         ];
-        $writer = new Writer($this->client, new NullLogger());
+        $writer = new TableWriter($this->client, new NullLogger());
         $tableQueue =  $writer->uploadTables($root . "/upload", $config, ["componentId" => "testComponent"]);
         $jobIds = $tableQueue->waitForAll();
         $this->assertCount(1, $jobIds);
@@ -331,7 +331,7 @@ class StorageApiWriterMetadataTest extends \PHPUnit_Framework_TestCase
                 ],
             ],
         ];
-        $writer = new Writer($this->client, new NullLogger());
+        $writer = new TableWriter($this->client, new NullLogger());
         $tableQueue =  $writer->uploadTables($root . "/upload", $config, ["componentId" => "testComponent"]);
         $jobIds = $tableQueue->waitForAll();
         $this->assertCount(1, $jobIds);
@@ -397,7 +397,7 @@ class StorageApiWriterMetadataTest extends \PHPUnit_Framework_TestCase
                 ],
             ],
         ];
-        $writer = new Writer($this->client, new NullLogger());
+        $writer = new TableWriter($this->client, new NullLogger());
         $tableQueue =  $writer->uploadTables($root . "/upload", $config, ["componentId" => "testComponent"]);
         $jobIds = $tableQueue->waitForAll();
         $this->assertCount(1, $jobIds);
@@ -471,7 +471,7 @@ class StorageApiWriterMetadataTest extends \PHPUnit_Framework_TestCase
                 ],
             ],
         ];
-        $writer = new Writer($this->client, new NullLogger());
+        $writer = new TableWriter($this->client, new NullLogger());
         $tableQueue =  $writer->uploadTables($root . "/upload", $config, ["componentId" => "testComponent"]);
         $jobIds = $tableQueue->waitForAll();
         $this->assertCount(1, $jobIds);
@@ -551,7 +551,7 @@ class StorageApiWriterMetadataTest extends \PHPUnit_Framework_TestCase
             "configurationRowId" => "row-1"
         ];
 
-        $writer = new Writer($this->client, new NullLogger());
+        $writer = new TableWriter($this->client, new NullLogger());
         $tableQueue =  $writer->uploadTables($root . "/upload", $config, $systemMetadata);
         $jobIds = $tableQueue->waitForAll();
         $this->assertCount(1, $jobIds);
